requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
click>=8.1.7
rich>=13.7.0
mysql-connector-python>=8.3.0
psycopg2-binary>=2.9.9
flask>=3.0.0
gunicorn>=22.0.0
